Adds a benchmark for FromUAPI in wgcfg.
Updates #1912.

I was also going to see if there were any easy perf gains here, but it doesn't appear that there's anything I can easily change. I should note there is one small change for `"replace_allowed_ips"` which wasn't handled before, and I'm not sure what indicates.