@JulianKnodt 2021-07-13
Ah ok now there are easy perf gains:
```
name        old time/op    new time/op    delta
FromUAPI-8    4.51µs ± 2%    3.47µs ± 8%  -23.15%  (p=0.008 n=5+5)

name        old alloc/op   new alloc/op   delta
FromUAPI-8    7.18kB ± 0%    6.64kB ± 0%   -7.57%  (p=0.008 n=5+5)

name        old allocs/op  new allocs/op  delta
FromUAPI-8      53.0 ± 0%      36.0 ± 0%  -32.08%  (p=0.008 n=5+5)
```