@JulianKnodt 2021-07-13
Latest change has a lot of perf improvements
```
name        old time/op    new time/op    delta
FromUAPI-8    4.51µs ± 2%    2.84µs ± 1%  -37.08%  (p=0.008 n=5+5)

name        old alloc/op   new alloc/op   delta
FromUAPI-8    7.18kB ± 0%    5.54kB ± 0%  -22.94%  (p=0.008 n=5+5)

name        old allocs/op  new allocs/op  delta
FromUAPI-8      53.0 ± 0%      13.0 ± 0%  -75.47%  (p=0.008 n=5+5)
```
Probably worth another review since there's a lot of changes, but if you don't think it's necessary it's good to merge now