@josharian 2021-07-13
I'm not sure this should have been merged. We have gone from fairly clear, and very similar to upstream code, to code that is a lot less nice.

Aside from microbenchmark improvements, what do we gain from this? I don't recall seeing these show up in pprof anywhere.