@JulianKnodt 2021-07-13
If you feel strongly about it, then I can revert it, also since I realize this is around the release timing. That being said, I think all of the changes in this PR are strictly positive and adhere closely to the original structure but with some type modifications, and the change in readability is worth having it.