@bradfitz 2021-07-13
Before reverting, let's look at hello's profiles.

Sorry, I'd assumed this was motivated by something in particular.