@bradfitz 2021-07-13
Looking at pprof --alloc_space ...

```
      12MB 0.00021%   100%  1960.54MB 0.035%  tailscale.com/wgengine/wgcfg.FromUAPI
```

And focusing on UAPI:

<img width="1249" alt="Screen Shot 2021-07-13 at 2 41 13 PM" src="https://user-images.githubusercontent.com/2621/125529009-48074ffb-c321-4f13-ba96-cbc64d089cbb.png">