@josharian 2021-07-13
Wow, hello allocates a lot, if 1.91GB is only 0.035% of its allocations.

Anyway, 0.035% doesn't seem like enough to be worth optimizing for.

Of course, as a percentage of allocs, that'll go up once the package net optimizations to avoid per-packet allocations are in. But I still suspect it'll still be fairly irrelevant overall; I suspect that at that point, we won't be allocating enough for the allocations to matter. Happy to be proved wrong, though.